# Faster IPC write