```
write i64 2^20          time:   [6.0850 ms 6.1489 ms 6.2148 ms]                           
                        change: [-88.115% -87.932% -87.756%] (p = 0.00 < 0.05)

write bool 2^20         time:   [23.463 us 23.651 us 23.838 us]                             
                        change: [-32.607% -32.010% -31.422%] (p = 0.00 < 0.05)

write utf8 2^20         time:   [1.5520 ms 1.5613 ms 1.5710 ms]                             
                        change: [-98.498% -98.483% -98.467%] (p = 0.00 < 0.05)
```

This also reduces the size of the file for sliced arrays by only writing required values (e.g. not write values that have been sliced away).

Closes #192